+++ moqt/session_test.go
@@ -48,9 +48,13 @@ func (noStatsConn) CloseWithError(transport.ConnErrorCode, string) error { retur
 func (noStatsConn) Context() context.Context                             { return context.Background() }
 func (noStatsConn) LocalAddr() net.Addr                                  { return &net.TCPAddr{} }
 func (noStatsConn) OpenStream() (transport.Stream, error)                { return nil, io.EOF }
+func (noStatsConn) OpenStreamSync(context.Context) (transport.Stream, error) { return nil, io.EOF }
 func (noStatsConn) OpenUniStream() (transport.SendStream, error)         { return nil, io.EOF }
-func (noStatsConn) RemoteAddr() net.Addr                                 { return &net.TCPAddr{} }
-func (noStatsConn) TLS() *tls.ConnectionState                            { return nil }
+func (noStatsConn) OpenUniStreamSync(context.Context) (transport.SendStream, error) {
+	return nil, io.EOF
+}
+func (noStatsConn) RemoteAddr() net.Addr      { return &net.TCPAddr{} }
+func (noStatsConn) TLS() *tls.ConnectionState { return nil }
 
 func TestNewSession(t *testing.T) {
 	tests := map[string]struct {